requirements.txt
pandas==2.2.2
psycopg[binary]==3.2.9
python-dotenv==1.0.1
pytest==8.2.2